Add files via upload
**Writing the code from scratch

Because I couldn't tell which way was up on the previous version. I've decided to start anew.
In this version the layout will more or less be the same I'm guessing. But the entire JS code will be new.
I will try and make it so I can add future updates. This means logically building my program with different functions that I can expand upon. 

Wish me luck